backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
django-sendgrid-v5==1.2.4
mysqlclient==2.2.1
